satisfied code review

+++ kernel/include/rendering.h
@@ -17,14 +17,15 @@ typedef struct Framebuffer {
     FBFormat format;
 } __attribute__((packed)) Framebuffer;
 
-extern uint64_t g_bg_color;
-extern uint64_t g_fg_color;
+extern uint32_t g_bg_color;
+extern uint32_t g_fg_color;
 extern Framebuffer* g_frame_buffer;
 
 void put_pixel(uint64_t x, uint64_t y, uint32_t color);
 void put_char(char c, uint64_t x, uint64_t y, uint32_t fg, uint32_t bg);
 
-//print value in hex, where size is determined and returned
+//print value in hex
+//returns the number of characters written (including 0x)
 uint64_t print_hex(uint64_t value, uint64_t x, uint64_t y);
 
 //print a certain amount of hex characters
@@ -37,7 +38,7 @@ void print_hex_32(uint64_t value, uint64_t x, uint64_t y);
 void print_hex_64(uint64_t value, uint64_t x, uint64_t y);
 
 //clear screen with a certain color
-void clear_screen(uint64_t color);
+void clear_screen(uint32_t color);
 
 //print string and return length
 uint64_t print_string(char* str, uint64_t x, uint64_t y);
